+++ code/kmeans_omp.cpp
@@ -1,7 +1,7 @@
 #include "kmeans.h"
 #include "kmeans_omp.h"
 
-int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize) {
+int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize) {
     reset_timer( 1 );
     reset_timer( 2 );
     reset_timer( 3 );
@@ -29,16 +29,16 @@ int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dim
         for (int particle_iter = 0; particle_iter < particle_count; particle_iter++) { // TODO parallel for
             // find closest center
             int cluster_assignment = (int)assignments[particle_iter];
-            float min_dist;
+            double min_dist;
             if(cluster_assignment < cluster_count) {
-                min_dist = compute_distance<uchar, float>(particle_data, particle_iter, centers, cluster_assignment, dimensions);
+                min_dist = compute_distance<uchar, double>(particle_data, particle_iter, centers, cluster_assignment, dimensions);
             }
             else {
-                min_dist = std::numeric_limits<float>::max();
+                min_dist = std::numeric_limits<double>::max();
             }
             for (int center_iter = 0; center_iter < cluster_count; center_iter++) {
                 if( center_iter != cluster_assignment ) {
-                    float dist = compute_distance<uchar, float>(particle_data, particle_iter, centers, center_iter, dimensions);
+                    double dist = compute_distance<uchar, double>(particle_data, particle_iter, centers, center_iter, dimensions);
                     if (dist < min_dist) {
                         min_dist = dist;
                         cluster_assignment = center_iter;
@@ -60,7 +60,7 @@ int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dim
             // initialize centers to 0
             for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {
                 for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                    array_store<float>(centers, cluster_iter, dim_iter, dimensions) = 0;
+                    array_store<double>(centers, cluster_iter, dim_iter, dimensions) = 0;
                 }
             }
             for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {
@@ -72,7 +72,7 @@ int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dim
                 int cluster_assignment = (int)assignments[particle_iter];
                 cluster_sizes[cluster_assignment]++;
                 for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                    array_store<float>(centers, cluster_assignment, dim_iter, dimensions) += (float)array_load<uchar>(particle_data, particle_iter, dim_iter, dimensions);
+                    array_store<double>(centers, cluster_assignment, dim_iter, dimensions) += (double)array_load<uchar>(particle_data, particle_iter, dim_iter, dimensions);
 
                 }
             }
@@ -81,8 +81,8 @@ int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dim
             #pragma omp parallel for
             for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {
                 for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                    array_store<float>(centers, cluster_iter, dim_iter, dimensions) =
-                        floor(array_load<float>(centers, cluster_iter, dim_iter, dimensions)/cluster_sizes[cluster_iter] + 0.5f);
+                    array_store<double>(centers, cluster_iter, dim_iter, dimensions) =
+                        floor(array_load<double>(centers, cluster_iter, dim_iter, dimensions)/cluster_sizes[cluster_iter] + 0.5f);
                 }
             }
             #pragma omp barrier
@@ -105,8 +105,8 @@ int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dim
 /**
  * Smarter cluster select method
  */
-int select_centerspp_omp(uchar *particle_data, int particle_count, int dimensions, int cluster_count, float *centers) {
-    float *particle_distr = new float[particle_count];
+int select_centerspp_omp(uchar *particle_data, int particle_count, int dimensions, int cluster_count, double *centers) {
+    double *particle_distr = new double[particle_count];
     int distr_length = RAND_MAX;
 
     reset_timer(4);
@@ -117,7 +117,7 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
     // Select the first center randomly
     int particle_select = rand() % particle_count;
     for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-        array_store<float>(centers, 0, dim_iter, dimensions) = (float)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
+        array_store<double>(centers, 0, dim_iter, dimensions) = (double)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
     }
 
     // Select the other n - 1 centers
@@ -127,9 +127,9 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
         #pragma omp parallel for
         for (int particle_iter = 0; particle_iter < particle_count; particle_iter++) { // TODO parallel for
             // find closest center based on previously calculated centers
-            float min_dist = std::numeric_limits<float>::max();
+            double min_dist = std::numeric_limits<double>::max();
             for (int center_dist_iter = 0; center_dist_iter < center_iter; center_dist_iter++) {
-                float dist = compute_distance<uchar, float>(particle_data, particle_iter, centers, center_dist_iter, dimensions);
+                double dist = compute_distance<uchar, double>(particle_data, particle_iter, centers, center_dist_iter, dimensions);
                 if (dist < min_dist) {
                     min_dist = dist;
                 }
@@ -157,7 +157,7 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
 
         start_timer(6);
         // Select new center
-        float distr_ratio = particle_distr[particle_count-1] / distr_length;
+        double distr_ratio = particle_distr[particle_count-1] / distr_length;
         int select_pos = rand() % distr_length;
         #pragma omp parallel for
         for (int particle_iter = 0; particle_iter < particle_count; particle_iter++) {  // TODO search - make binary, parallelize (note break, tricky)
@@ -170,7 +170,7 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
 
         // Copy new center to array
         for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-            array_store<float>(centers, center_iter, dim_iter, dimensions) = (float)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
+            array_store<double>(centers, center_iter, dim_iter, dimensions) = (double)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
         }
     }
     delete [] particle_distr;
